app/requirements.txt
python-dotenv==1.1.0
beautifulsoup4==4.13.4
playwright==1.52.0
requests==2.32.3